requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
astroid==2.4.2
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
blinker==1.4
blueprint==3.4.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
discord-pretty-help==1.3.2
discord.py==1.7.3
easypydb==0.4.4
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Discord==0.1.64
Flask-SQLAlchemy==2.5.1
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-spreadsheet==0.0.6
googleapis-common-protos==1.52.0
greenlet==1.1.0
gspread==3.7.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httplib2==0.18.1
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
importlib-metadata==4.6.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.2.0
priority==2.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
PyJWT==2.1.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
Quart==0.15.1
Quart-Discord==2.1.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
s1db==1.0.3
six==1.15.0
slackclient==2.9.3
SQLAlchemy==1.4.22
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==2.0.1
wrapt==1.12.1
wsproto==1.0.0
WTForms==2.3.3
yarl==1.6.3
zipp==3.5.0
dhooks==1.1.4